kalman_filter_test now correctly passes in a dictionary of inputs and returns a dictonary of outputs. Print array utility functions working as intended. Correct dimensions of coerced / copied inputs & outputs. Staged returning of actual outputs.

+++ test.h
@@ -18,4 +18,4 @@ double **ckalman_filter_test(
     double *HHt, int incHHt,
     double *GGt, int incGGt,
     double *yt,
-    int *loglik);
+    double *loglik);
